backend/requirements.txt
flask
pycryptodome
redis
flask-cors

